requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-telegram-bot==21.*
openai==1.*
APScheduler==3.10.4
pytz
google-generativeai
aiogram
requests
sqlalchemy
telegram
werkzeug